+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] / 2023-02-13
+### Features
+- Feature `TestEngine.GetTestFullNames` with `directoryResolver`
+- Feature `AssemblyResolveService` improved
+### Added
+- Add new `AssemblyResolveService`
+- Add `InternalsVisibleTo` Tests
+### Updated
+- Update all services to internal
+- Update `AssemblyResolveService` to `AssemblyResolveCurrentDirectoryService`
+
 ## [1.1.1] / 2022-12-21
 ### Features
 - Feature `TestModel` with FullName